CAUSEWAY-3815: flatten ComponentFactoryRegistry type hierarchy
