+++ viewers/wicket/ui/src/main/java/org/apache/causeway/viewer/wicket/ui/app/registry/ComponentFactoryRegistry.java
@@ -18,8 +18,11 @@
  */
 package org.apache.causeway.viewer.wicket.ui.app.registry;
 
+import java.util.HashMap;
+import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.Optional;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import org.apache.wicket.Component;
@@ -29,8 +32,16 @@ import org.apache.wicket.model.IModel;
 import org.springframework.lang.Nullable;
 
 import org.apache.causeway.commons.collections.ImmutableEnumSet;
+import org.apache.causeway.commons.internal.base._Refs;
+import org.apache.causeway.commons.internal.base._Text;
+import org.apache.causeway.commons.internal.collections._Multimaps;
+import org.apache.causeway.commons.internal.collections._Multimaps.ListMultimap;
+import org.apache.causeway.core.metamodel.context.MetaModelContext;
 import org.apache.causeway.viewer.commons.model.components.UiComponentType;
 import org.apache.causeway.viewer.wicket.ui.ComponentFactory;
+import org.apache.causeway.viewer.wicket.ui.ComponentFactoryAbstract;
+
+import lombok.extern.log4j.Log4j2;
 
 /**
  * API for finding registered {@link ComponentFactory}s.
@@ -38,21 +49,28 @@ import org.apache.causeway.viewer.wicket.ui.ComponentFactory;
  * Ultimately all requests to locate {@link ComponentFactory}s are routed
  * through to an object implementing this interface.
  */
-public interface ComponentFactoryRegistry {
+@Log4j2
+public final class ComponentFactoryRegistry {
 
-    Stream<ComponentFactory> streamComponentFactories(
-            UiComponentType uiComponentType, @Nullable IModel<?> model);
+    private final ListMultimap<UiComponentType, ComponentFactory> componentFactoriesByComponentType =
+            _Multimaps.newListMultimap();
+    private final Map<Class<? extends ComponentFactory>, ComponentFactory> componentFactoriesByType =
+            new HashMap<>();
 
-    Stream<ComponentFactory> streamComponentFactories(
-            ImmutableEnumSet<UiComponentType> uiComponentTypes,
-            @Nullable IModel<?> model);
+    public ComponentFactoryRegistry(
+            final ComponentFactoryList factoryList,
+            final MetaModelContext mmc) {
+        super();
+        factoryList.forEach(compFactory->this.registerComponentFactory(mmc, compFactory));
+        ensureAllComponentTypesRegistered();
+    }
 
     /**
      * Finds the "best" {@link ComponentFactory} for given componentType.
      * <p>
      * Falls back to a {@link UiComponentType#UNKNOWN} lookup.
      */
-    default ComponentFactory findComponentFactory(
+    public ComponentFactory findComponentFactory(
             final UiComponentType uiComponentType, final @Nullable IModel<?> model) {
         return streamComponentFactories(uiComponentType, model)
             .findFirst()
@@ -61,7 +79,7 @@ public interface ComponentFactoryRegistry {
                     .orElse(null));
     }
 
-    default ComponentFactory findComponentFactoryElseFail(
+    public ComponentFactory findComponentFactoryElseFail(
             final UiComponentType uiComponentType, final @Nullable IModel<?> model) {
         return streamComponentFactories(uiComponentType, model)
                 .findFirst()
@@ -77,17 +95,24 @@ public interface ComponentFactoryRegistry {
      * {@link #addOrReplaceComponent(MarkupContainer, UiComponentType, IModel)},
      * but with the wicket id derived from the {@link UiComponentType}.
      */
-    Component addOrReplaceComponent(MarkupContainer markupContainer, UiComponentType uiComponentType, IModel<?> model);
+    public Component addOrReplaceComponent(final MarkupContainer markupContainer, final UiComponentType uiComponentType, final IModel<?> model) {
+        final Component component = createComponent(uiComponentType, model);
+        markupContainer.addOrReplace(component);
+        return component;
+    }
 
     /**
      * {@link #createComponent(String, UiComponentType, IModel) Creates} the
      * relevant {@link Component} for the provided arguments, and adds to the
      * provided {@link MarkupContainer}; the wicket id is as specified.
-     *
      * <p>
      * If none can be found, will fail fast.
      */
-    Component addOrReplaceComponent(MarkupContainer markupContainer, String id, UiComponentType uiComponentType, IModel<?> model);
+    public Component addOrReplaceComponent(final MarkupContainer markupContainer, final String id, final UiComponentType uiComponentType, final IModel<?> model) {
+        final Component component = createComponent(id, uiComponentType, model);
+        markupContainer.addOrReplace(component);
+        return component;
+    }
 
     /**
      * As per {@link #createComponent(String, UiComponentType, IModel)}, but with
@@ -95,7 +120,10 @@ public interface ComponentFactoryRegistry {
      *
      * @see #createComponent(String, UiComponentType, IModel)
      */
-    Component createComponent(UiComponentType uiComponentType, IModel<?> model);
+    public Component createComponent(final UiComponentType uiComponentType, final IModel<?> model) {
+        return findComponentFactoryElseFail(uiComponentType, model)
+                .createComponent(model);
+    }
 
     /**
      * Create the {@link Component} matching the specified {@link UiComponentType}
@@ -105,13 +133,97 @@ public interface ComponentFactoryRegistry {
      * <p>
      * If none can be found, will fail fast.
      */
-    Component createComponent(String id, UiComponentType uiComponentType, IModel<?> model);
+    public Component createComponent(final String id, final UiComponentType uiComponentType, final IModel<?> model) {
+        return findComponentFactoryElseFail(uiComponentType, model)
+                .createComponent(id, model);
+    }
 
-    <T extends ComponentFactory> Optional<T> lookupFactory(Class<T> factoryClass);
-    default <T extends ComponentFactory> T lookupFactoryElseFail(final Class<T> factoryClass) {
+    public Stream<ComponentFactory> streamComponentFactories(
+            final UiComponentType uiComponentType,
+            final @Nullable IModel<?> model) {
+
+        // find all that apply, unless we find one that applies exclusively
+        // in the exclusive case, we just return the exclusive one
+
+        var exclusiveIfAny = _Refs.<ComponentFactory>objectRef(null);
+
+        var allThatApply = componentFactoriesByComponentType.streamElements(uiComponentType)
+                .filter(componentFactory->{
+                    var advice = componentFactory.appliesTo(uiComponentType, model);
+                    if(advice.appliesExclusively()) {
+                        exclusiveIfAny.set(componentFactory);
+                    }
+                    return advice.applies();
+                })
+                // as an optimization, stop taking when we found an exclusive one
+                .takeWhile(__->exclusiveIfAny.isNull())
+                .collect(Collectors.toList());
+
+        return (exclusiveIfAny.isNotNull()
+                    ? Stream.of(exclusiveIfAny.getValueElseFail())
+                    : allThatApply.stream()
+                )
+                .peek(componentFactory->logComponentResolving(model, uiComponentType, componentFactory));
+    }
+
+    public Stream<ComponentFactory> streamComponentFactories(
+            final ImmutableEnumSet<UiComponentType> uiComponentTypes,
+            final @Nullable IModel<?> model) {
+        return uiComponentTypes.stream()
+                .flatMap(componentType->streamComponentFactories(componentType, model));
+    }
+
+    @SuppressWarnings("unchecked")
+    public <T extends ComponentFactory> Optional<T> lookupFactory(final Class<T> factoryClass) {
+        return Optional.ofNullable((T)componentFactoriesByType.get(factoryClass));
+    }
+
+    public <T extends ComponentFactory> T lookupFactoryElseFail(final Class<T> factoryClass) {
         return lookupFactory(factoryClass)
             .orElseThrow(()->
                 new NoSuchElementException("Could not locate component factory of type '" + factoryClass + "'"));
+    }
+
+    // -- HELPER
+
+    private void registerComponentFactory(
+            final MetaModelContext commonContext,
+            final ComponentFactory componentFactory) {
+        componentFactoriesByType.put(componentFactory.getClass(), componentFactory);
+
+        // handle dependency injection for factories
+        commonContext.getServiceInjector().injectServicesInto(componentFactory);
+        if(componentFactory instanceof ComponentFactoryAbstract) {
+            ((ComponentFactoryAbstract)componentFactory).setMetaModelContext(commonContext);
+        }
+
+        componentFactoriesByComponentType.putElement(componentFactory.getComponentType(), componentFactory);
+    }
+
+    private void ensureAllComponentTypesRegistered() {
+        for (var componentType : UiComponentType.values()) {
+
+            if(componentType.getOptionality().isOptional()) {
+                continue;
+            }
+
+            if (componentFactoriesByComponentType.getOrElseEmpty(componentType).isEmpty()) {
+                throw new IllegalStateException("No component factories registered for " + componentType);
+            }
+        }
+    }
+
+    // -- DEBUG LOGGING
+
+    private static void logComponentResolving(
+            final IModel<?> model,
+            final UiComponentType uiComponentType,
+            final ComponentFactory componentFactory) {
+        if(!log.isDebugEnabled()) return;
+        log.debug("component type for model {} -> {} provided by {}",
+                _Text.abbreviateClassOf(model),
+                uiComponentType.name(),
+                _Text.abbreviateClassOf(componentFactory));
     }
 
 }
